+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [10.1.0] - 2025-09-0?
 
+- Added: Action to abandon a world in a multiworld session.
+
 ### Metroid Dread
 
 #### Logic Database